refactor: move program creation to separate file for easier testing
